TSK: move actions to separate directory in app generator
