+++ generators/create-app/templates/cra-redux/src/actions/sampleApp/constants.js
