+++ generators/create-app/templates/cra-redux/src/actions/sampleApp/sampleActions.js
