+++ generators/create-app/templates/cra-redux/src/reducers/root/index.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux'
 import { routerReducer } from 'react-router-redux'
-import sampleReducer from '../sample/sampleReducer'
+import sampleReducer from '../sampleApp/sampleReducer'
 
 export default combineReducers({
   routing: routerReducer,
